week1/requirements.txt
PyMuPDF==1.23.1
openai==1.33.0
tiktoken==0.4.0
rich==13.6.0
python-dotenv==1.0.0